requirements.web.txt
alembic==1.17.2
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cx_Oracle==8.3.0
dnspython==2.8.0
email-validator==2.3.0
filelock==3.20.0
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fsspec==2025.12.0
greenlet==3.2.4
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
Mako==1.3.10
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
SQLAlchemy==2.0.21
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
Werkzeug==3.1.3
WTForms==3.2.1
xgboost==3.1.1
mcp
langgraph
langchain-huggingface
langchain-chroma
langchain-core
sentence-transformers